+++ cookbooks/iptables-nativex/attributes/privoxy.rb
@@ -1,2 +1,2 @@
 # Privoxy Allowed Networks (IP/NETMASK)
-default['nativex-iptables']['privoxy']['addresses'] = Array.new
+default['iptables-nativex']['privoxy']['addresses'] = Array.new