+++ cookbooks/iptables-nativex/attributes/ssh.rb
@@ -1,2 +1,2 @@
 # SSH Allowed Networks (IP/NETMASK)
-default['nativex-iptables']['ssh']['addresses'] = Array.new
+default['iptables-nativex']['ssh']['addresses'] = Array.new