+++ cookbooks/iptables-nativex/attributes/web.rb
@@ -1,2 +1,2 @@
 # Web Traffic Allowed Networks (IP/NETMASK)
-default['nativex-iptables']['web']['addresses'] = Array.new
+default['iptables-nativex']['web']['addresses'] = Array.new