+++ cookbooks/iptables-nativex/recipes/web.rb
@@ -1,5 +1,5 @@
 #
-# Cookbook Name:: nativex-iptables
+# Cookbook Name:: iptables-nativex
 # Recipe:: web.rb
 #
 # Copyright 2014, NativeX
@@ -8,7 +8,7 @@
 #
 
 # Have we decided to lock down the node?
-if node['nativex-iptables']['web']['addresses'].empty?
+if node['iptables-nativex']['web']['addresses'].empty?
   # Use the all_ rules
   iptables_rule "all_http"
   iptables_rule "all_https"