+++ cookbooks/iptables-nativex/CHANGELOG.md
@@ -1,11 +1,11 @@
-nativex-iptables CHANGELOG
+iptables-nativex CHANGELOG
 ==========================
 
-This file is used to list changes made in each version of the nativex-iptables cookbook.
+This file is used to list changes made in each version of the iptables-nativex cookbook.
 
 0.1.0
 -----
-- [your_name] - Initial release of nativex-iptables
+- [your_name] - Initial release of iptables-nativex
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.